pyproject.toml
[project]
name = "raven"
version = "0.1.0"
description = "Zero-budget, production-grade RAG system with agentic planning, hybrid retrieval, and multimodal capabilities"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    # Document Processing
    "docling>=2.0.0",
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.24.0",
    "python-magic>=0.4.27",

    # Vector & Embeddings
    "chromadb>=1.0.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",

    # LLM & Orchestration
    "ollama>=0.4.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",

    # Retrieval & Reranking
    "transformers>=4.40.0",
    "torch>=2.0.0",

    # Evaluation
    "ragas>=0.2.0",
    "datasets>=3.0.0",
    "scikit-learn>=1.5.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Frontend
    "streamlit>=1.38.0",
    "plotly>=5.20.0",

    # Storage & Caching
    "diskcache>=5.6.0",
    "aiosqlite>=0.20.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Search Fallback
    "duckduckgo-search>=6.0.0",

    # Logging
    "structlog>=24.0.0",

    # Config
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Graph
    "networkx>=3.0",
]

[project.optional-dependencies]
security = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
]
multimodal = [
    "open-clip-torch>=2.24.0",
    "openai-whisper>=20231117",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
